+++ test/hexapdf/content/test_canvas.rb
@@ -689,6 +689,22 @@ describe HexaPDF::Content::Canvas do
     end
   end
 
+  describe "line_with_rounded_corner" do
+    it "serializes correctly" do
+      @canvas.move_to(10, 10)
+      @canvas.line_with_rounded_corner(30, 10, 30, 50, in_radius: 10, out_radius: 5)
+      assert_operators(@canvas.contents,
+                       [[:move_to, [10, 10]],
+                        [:line_to, [20, 10]],
+                        [:curve_to, [25.51915, 10.0, 30.0, 12.240425, 30.0, 15.0]]])
+    end
+
+    it "returns the canvas object" do
+      @canvas.move_to(10, 10)
+      assert_equal(@canvas, @canvas.line_with_rounded_corner(30, 30, 30, 50, in_radius: 10))
+    end
+  end
+
   describe "graphic_object" do
     it "returns a new graphic object given a name" do
       arc = @canvas.graphic_object(:arc)